Fixed bug where specifying no aliases would cause an error.
Not specifying an alias for a switch would cause the code mapping the aliases
to the correct switches to think a duplicate switch had been introduced,
because the switch spec itself is a key in the same hash.

New tests added to reproduce this bug.

Thanks to Doug Seay for finding this and bringing it to my attention.